refactor: move DfScalarFunc to separate file
