+++ src/flow/src/expr/scalar.rs
@@ -42,6 +42,7 @@ use crate::expr::error::{
     InvalidArgumentSnafu, OptimizeSnafu, TypeMismatchSnafu,
 };
 use crate::expr::func::{BinaryFunc, UnaryFunc, UnmaterializableFunc, VariadicFunc};
+use crate::expr::DfScalarFunction;
 use crate::repr::{ColumnType, RelationDesc, RelationType};
 use crate::transform::{from_scalar_fn_to_df_fn_impl, FunctionExtensions};
 /// A scalar expression with a known type.
@@ -173,162 +174,6 @@ pub enum ScalarExpr {
         then: Box<ScalarExpr>,
         els: Box<ScalarExpr>,
     },
-}
-
-/// A way to represent a scalar function that is implemented in Datafusion
-#[derive(Debug, Clone)]
-pub struct DfScalarFunction {
-    raw_fn: RawDfScalarFn,
-    // TODO(discord9): directly from datafusion expr
-    fn_impl: Arc<dyn PhysicalExpr>,
-    df_schema: Arc<datafusion_common::DFSchema>,
-}
-
-impl DfScalarFunction {
-    pub fn new(raw_fn: RawDfScalarFn, fn_impl: Arc<dyn PhysicalExpr>) -> Result<Self, Error> {
-        Ok(Self {
-            df_schema: Arc::new(raw_fn.input_schema.to_df_schema()?),
-            raw_fn,
-            fn_impl,
-        })
-    }
-
-    pub async fn try_from_raw_fn(raw_fn: RawDfScalarFn) -> Result<Self, Error> {
-        Ok(Self {
-            fn_impl: raw_fn.get_fn_impl().await?,
-            df_schema: Arc::new(raw_fn.input_schema.to_df_schema()?),
-            raw_fn,
-        })
-    }
-
-    /// eval a list of expressions using input values
-    fn eval_args(values: &[Value], exprs: &[ScalarExpr]) -> Result<Vec<Value>, EvalError> {
-        exprs
-            .iter()
-            .map(|expr| expr.eval(values))
-            .collect::<Result<_, _>>()
-    }
-
-    // TODO(discord9): add RecordBatch support
-    pub fn eval(&self, values: &[Value], exprs: &[ScalarExpr]) -> Result<Value, EvalError> {
-        // first eval exprs to construct values to feed to datafusion
-        let values: Vec<_> = Self::eval_args(values, exprs)?;
-        if values.is_empty() {
-            return InvalidArgumentSnafu {
-                reason: "values is empty".to_string(),
-            }
-            .fail();
-        }
-        // TODO(discord9): make cols all array length of one
-        let mut cols = vec![];
-        for (idx, typ) in self
-            .raw_fn
-            .input_schema
-            .typ()
-            .column_types
-            .iter()
-            .enumerate()
-        {
-            let typ = typ.scalar_type();
-            let mut array = typ.create_mutable_vector(1);
-            array.push_value_ref(values[idx].as_value_ref());
-            cols.push(array.to_vector().to_arrow_array());
-        }
-        let schema = self.df_schema.inner().clone();
-        let rb = DfRecordBatch::try_new(schema, cols).map_err(|err| {
-            ArrowSnafu {
-                context:
-                    "Failed to create RecordBatch from values when eval datafusion scalar function",
-            }
-            .into_error(err)
-        })?;
-
-        let res = self.fn_impl.evaluate(&rb).map_err(|err| {
-            EvalDatafusionSnafu {
-                raw: err,
-                context: "Failed to evaluate datafusion scalar function",
-            }
-            .build()
-        })?;
-        let res = common_query::columnar_value::ColumnarValue::try_from(&res)
-            .map_err(BoxedError::new)
-            .context(ExternalSnafu)?;
-        let res_vec = res
-            .try_into_vector(1)
-            .map_err(BoxedError::new)
-            .context(ExternalSnafu)?;
-        let res_val = res_vec
-            .try_get(0)
-            .map_err(BoxedError::new)
-            .context(ExternalSnafu)?;
-        Ok(res_val)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct RawDfScalarFn {
-    /// The raw bytes encoded datafusion scalar function
-    pub(crate) f: bytes::BytesMut,
-    /// The input schema of the function
-    pub(crate) input_schema: RelationDesc,
-    /// Extension contains mapping from function reference to function name
-    pub(crate) extensions: FunctionExtensions,
-}
-
-impl RawDfScalarFn {
-    pub fn from_proto(
-        f: &substrait::substrait_proto_df::proto::expression::ScalarFunction,
-        input_schema: RelationDesc,
-        extensions: FunctionExtensions,
-    ) -> Result<Self, Error> {
-        let mut buf = BytesMut::new();
-        f.encode(&mut buf)
-            .context(EncodeRelSnafu)
-            .map_err(BoxedError::new)
-            .context(crate::error::ExternalSnafu)?;
-        Ok(Self {
-            f: buf,
-            input_schema,
-            extensions,
-        })
-    }
-    async fn get_fn_impl(&self) -> Result<Arc<dyn PhysicalExpr>, Error> {
-        let f = ScalarFunction::decode(&mut self.f.as_ref())
-            .context(DecodeRelSnafu)
-            .map_err(BoxedError::new)
-            .context(crate::error::ExternalSnafu)?;
-        debug!("Decoded scalar function: {:?}", f);
-
-        let input_schema = &self.input_schema;
-        let extensions = &self.extensions;
-
-        from_scalar_fn_to_df_fn_impl(&f, input_schema, extensions).await
-    }
-}
-
-impl std::cmp::PartialEq for DfScalarFunction {
-    fn eq(&self, other: &Self) -> bool {
-        self.raw_fn.eq(&other.raw_fn)
-    }
-}
-
-// can't derive Eq because of Arc<dyn PhysicalExpr> not eq, so implement it manually
-impl std::cmp::Eq for DfScalarFunction {}
-
-impl std::cmp::PartialOrd for DfScalarFunction {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-impl std::cmp::Ord for DfScalarFunction {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.raw_fn.cmp(&other.raw_fn)
-    }
-}
-impl std::hash::Hash for DfScalarFunction {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.raw_fn.hash(state);
-    }
 }
 
 impl ScalarExpr {
@@ -799,6 +644,7 @@ mod test {
     use substrait::substrait_proto_df::proto::{r#type, FunctionArgument, Type};
 
     use super::*;
+    use crate::expr::RawDfScalarFn;
     #[test]
     fn test_extract_bound() {
         let test_list: [(ScalarExpr, Result<_, EvalError>); 5] = [
@@ -887,39 +733,5 @@ mod test {
         let permute_map = BTreeMap::from([(1, 2), (3, 4)]);
         let res = expr.permute_map(&permute_map);
         assert!(matches!(res, Err(Error::InvalidQuery { .. })));
-    }
-
-    #[tokio::test]
-    async fn test_df_scalar_function() {
-        let raw_scalar_func = ScalarFunction {
-            function_reference: 0,
-            arguments: vec![FunctionArgument {
-                arg_type: Some(ArgType::Value(Expression {
-                    rex_type: Some(RexType::Literal(Literal {
-                        nullable: false,
-                        type_variation_reference: 0,
-                        literal_type: Some(LiteralType::I64(-1)),
-                    })),
-                })),
-            }],
-            output_type: None,
-            ..Default::default()
-        };
-        let input_schema = RelationDesc::try_new(
-            RelationType::new(vec![ColumnType::new_nullable(
-                ConcreteDataType::null_datatype(),
-            )]),
-            vec!["null_column".to_string()],
-        )
-        .unwrap();
-        let extensions = FunctionExtensions::from_iter(vec![(0, "abs")]);
-        let raw_fn = RawDfScalarFn::from_proto(&raw_scalar_func, input_schema, extensions).unwrap();
-        let df_func = DfScalarFunction::try_from_raw_fn(raw_fn).await.unwrap();
-        assert_eq!(
-            df_func
-                .eval(&[Value::Null], &[ScalarExpr::Column(0)])
-                .unwrap(),
-            Value::Int64(1)
-        );
     }
 }
